rename Drawer to PrimitiveDrawer
