+++ packages/design-system/src/components/Drawer/Primitive/PrimitiveDrawer.scss
@@ -1,6 +1,6 @@
 @use '~@talend/design-tokens/lib/tokens';
 
-.drawer {
+.primitive-drawer {
 	width: 30rem;
 	height: 100%;
 	background: tokens.$coral-color-neutral-background;
