+++ packages/design-system/src/components/Drawer/index.ts
@@ -1,4 +1,3 @@
 import { FloatingDrawer } from './variants/FloatingDrawer/FloatingDrawer';
 
 export { FloatingDrawer };
-export default FloatingDrawer;
